SignalTransition: guard against deleted object
If the object has been deleted, we can not retrieve the sender and
signalIndex anymore. In that case, the only thing we can do is to print
a warning.
This is impossible to trigger during normal usage from C++, however
it can happen in QML, where we do not disconnect handlers soon enough.
This issue only exists in Qt 5, as in Qt 6
c249edb83fa67b3e5f711b28923397e66876182d in qtdeclarative and a few
other changes take care of it.
A testcase for this issue will be added to qtdeclarative.

Fixes: QTBUG-65926
Change-Id: I694a77672c311039c4fe238d9e24c91a427e85e3
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>