Remove superfluous qfuture.h includes from qcoreapplication[_p].h
This include is only needed on Android, and removing it improves build
times slightly.

Remove the include from qcoreapplication.h and guard the includes in
qcoreapplication_platform.h properly for Android.

Remove the include from qcoreapplication_p.h as well.  This was a
left-over from the app permissions API that was removed.

Pick-to: 6.2
Task-number: QTBUG-97601
Change-Id: I131baebe3e08e93ad5420d40908c2dceab89554c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>